scripts/blog-pipeline/requirements.txt
scrapy>=2.11,<3.0
anthropic>=0.40,<1.0
httpx>=0.27,<1.0
supabase>=2.0,<3.0
pyyaml>=6.0,<7.0
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0
click>=8.0,<9.0
beautifulsoup4>=4.12,<5.0
python-dotenv>=1.0,<2.0
